ENH: Initial commit on plugin/dispatcher feature.
Added Dispatcher class and tests, allowing for modular plugin interface
to be defined by configuration file and loaded at runtime.
The idea is to allow plugins to be dropped in and run on specific
events, defining a basic interface that the plugins must conform to,
then dynamically loading them and running in a thread.